Make value objects hashable
So I can implement dictionaries

This might be a bit dicey, since they're *technically* mutable, but
the parts I mutate aren't involved in the hash, so it *might* be ok? I
know I'll run into issues if I try to hash things like arrays, but I'll
cross that bridge when I come to it.